Update - August 19, 2025
Large Fix - Found that in the DB1B Condense function that the Spirit online booking fare adjustment happened after initial stages of condensing process, rather than before this process. This has been fixed. Some values are different in the pre-pandemic results due to this and need to be corrected within the paper text.